2005-07-26  Atsushi Enomoto  <atsushi@ximian.com>
	* mono-tailoring-source.txt : Added all single sortkey remapping to
	  all cultures (still need to fill contractions and annotate possible
	  buggy mapping referencing to CLDR).
	* SimpleCollator.cs : removed unused code.
	* MSCompatUnicodeTable.cs : tiny cast removal.


svn path=/trunk/mcs/; revision=47689

+++ mcs/class/corlib/Mono.Globalization.Unicode/MSCompatUnicodeTable.cs
@@ -626,7 +626,7 @@ namespace Mono.Globalization.Unicode
 			idx += 4;
 			tailorings = new char [count];
 			for (int i = 0; i < count; i++, idx += 2)
-				tailorings [i] = (char) (ushort) (tailor [idx] + (tailor [idx + 1] << 8));
+				tailorings [i] = (char) (tailor [idx] + (tailor [idx + 1] << 8));
 			isReady = true;
 		}
